requirements.txt
django
beautifulsoup4
requests
lxml
pydantic
djangorestframework